[5.3] Create environment variables during installation (#213)
* Treat external resources as a collection of Kubernetes/Gravity
resources. Handle the ability to specify cluster runtime environment
variables resources.
Remove redundant resources package and attributes in cluster - instead
move resource handling into the install FSM.

* This adds additional logic to handling of external resources during
installation.
Before this change, the install command would only accept Kubernetes
resources as '--config' and apply them once the runtime container has
been installed.
With the change, we also handle Gravity resources as part of the file
specified in '--config'. The main motivation is being able to set
cluster runtime environment variables during installation to avoid
expensive update operation. But it made sense to extend support to other
Gravity resource types which are applied as the last step of the install
operation.

* Fix typos and comments

* Add the ability to append resources (as the case is with Ops Center
install when the process implicitly generates Ops Center configuration
as a ConfigMap to be added at some point into the operation).
Fix an issue with local custer operator not available at the start of
the install operation.

* Address review comments

* Remove debug logging and rephrase a doc-comment on storage.UnknownResource

* Update e to release