[5.4] 'gravity plan' consistency update and minor bug fixes throughout (#251)
Update startAgents in lib/ops/opsservice to start the agents and the update operation in automatic mode.
The agent started on one of the master nodes will be the driver for the operation. The same node will be used to initialize the operation plan (as the operation is started by the cluster controller in this
workflow).
Also, make agent deploy command line more explicit by adding explicit options to specify additional command parameters for leader and regular agents.

Remove 'gravity plan init' and 'gravity plan sync' as these were always specific to the update operation. Move 'gravity plan init' as 'gravity update init-plan' as this command is still used when triggering
update from UI.

Thread operation/operation ID to every function implementing the functionality of 'gravity plan' so it's possible to focus it explicitly on a specific operation.

Remove dns from endpoints wait loop as kube-dns/coreDNS changes to 5.3.x/5.4.x have broken the endpoints/service and the simplest workaround in these versions would be to ignore them completely.

Start update operation in blocking mode in CLI (unless overridden with the '--manual' flag) in foreground as opposed to running in background in a systemd unit as this is expected by the user and consistent with other cluster operation workflows.

* Pass operation to executeUpgradePhase when executed via 'gravity upgrade --phase'

* Update etcd phase sub-steps to depend on previous steps executed on the same node where necessary. This is to allow the remote agents to make progress during etcd upgrade even when they
cannot properly sync changes from upstream (leader) node.

* Add non-blocking operating mode for update (as before) to avoid breaking the robotest