Fix a flaky test (#511)
* Fix a flaky test:
When the test generates the application packages for local/remote services, there's a potential race if two separate packages are generated with the same contents.
The issue is with file/directory item timestamps that can make the two packages differ. Instead, generate a single application package and replicate it in both services.

* Address review comments

* Update Makefile to enable testing of all relevant packages.
Fix updateCommandWithFlags to properly handle arguments (use the value instead of their name).