Simplify commit determination.
Previously the code grepped through all tags, looking for a dereferenced
version of $SHORT_TAG. This fails when $SHORT_TAG is not an annotated
tag, but a lightweight tag (and thus has no dereference).

`git show-rev` will always find the commit a tag points to, regardless
of tag type, and is considerably shorter.