I've reconsidered the earlier decision to keep SELinux state cluster-wide and instead this commit make this node-specific.
The reasoning being that it is easier to imagine a heterogeneous cluster
where there might be only a subset of nodes using SELinux to provide
additional protection while the rest isn't. This is similar to an idea
of having nodes running disparate (albeit all supported) OS
distributions - i.e. to remove the 'same OS' restriction as there also
maybe be use-cases when various distribution might be good on nodes
running particular workloads.