Switch docbox user from jenkins -> builder.
This username is simply a placeholder for the actual user of the system.
It has nothing to do with jenkins. Having the name jenkins implies that
they're somehow related or interplay with each other... which is
misleading.

"builder" feels like "jenkins" (implying a user for build purposes) but
without the tie to a specific system.

I also took the opportunity to swap to bash instead of sh, as bash is
present, and provides a better UX.