Wait for operator. (#1285)
The previous attempt to fix the issue with waiting for process listeners to come up by embedding "wait" function into the process itself has some deficiencies as it turns out. In particular, process itself can't really reliably call it's own API in a general use-case b/c it doesn't know how clients are supposed to reach it: for example in installer use-case the advertise address is the IP address which works fine but when installing a hub, the "pack" address (which I used before) points to the public advertise hostname of the hub.

This PR instead moves the logic of checking the operator status to the caller - I integrated it into RemoteEnvironment instead which I think suits its purpose cause it knows the address of the installer it talks to, and the installer calls this method explicitly.