Add a flag to skip Jenkins parameter definition.
For long lived jobs (such as an unbranched gravity try build) we do not
want to re-define the job parameters each time the Jenkinsfile is
loaded.  This allows the job to be configured with a superset of the
parameters that may be defined in the Jenkinsfile -- which is necessary
when jobs need a GIT_BRANCH or similar parameter.