Fix jenkins config lag & try build support (#1327)
* Remove Robotest Jenkins parameters.

Allowing these values to be set via build paramaters results in
unexpected config lag (issue #1315) and is not currently by developers.

For simplicity, we allow these values to be overriden, but move the
defaults from Jenkinsfile to shell scripts.

Fixes #1315.

* Add a flag to skip Jenkins parameter definition.

For long lived jobs (such as an unbranched gravity try build) we do not
want to re-define the job parameters each time the Jenkinsfile is
loaded.  This allows the job to be configured with a superset of the
parameters that may be defined in the Jenkinsfile -- which is necessary
when jobs need a GIT_BRANCH or similar parameter.

* Remove unused AWS config.

Robotest does not currently run against AWS.  Instead of lingering half
config around, remove these values until such time as they're regularly
tested and working again.