Clarify end of support for non-LTS releases.
Instead of using a '-' to mark end of support, I've filled in the date
of the x.x.0 release that supersedes each unsupported minor version.

This is sometimes confusing as the release date for the most recent
patch release is after support ended.  However that is an accurate
representation of our development practices. On a case-by-cases basis
we will put out patch releases for paying customers, even after
official support for a release has ended.