Adjust upgrade instructions heading level.
The "Instructions on upgrading to 4.23.0+" heading was at the same level
as 4.x.  This resulted in all the 4.x releases nesting under
"Instructions..." and caused "Instructions..." to show up as in the ToC.

Perhaps it may have been important to highlight these upgrade
instructions at one point in the past, but we no longer support 4.x or
anything prior.  Thus a cleaner ToC hierarchy is probably more important
these days.