Increase precedence of CentOS:7 testing, pin 7.0.13.
After discussions with heavy users of Gravity 6.3.x and 7.0.x, we want
to do more CentOS:7 testing (and 8, but it isn't available without
https://github.com/gravitational/robotest/issues/258).

Thus we replace several ubuntu/redhat scenarios in PR builds with centos
versions of the same thing.

Furthermore, testing 7.0.13 upgrades with CentOS is particularly
important to our field usage.