Use helm to upgrade from echoserver 1.0 to 1.10.
Name the helm release such that we can refer back to it from the
postupdate hook. Add a postupdate hook to run `helm upgrade`.

Decrease upgrade-base echosever image to 1.0, such that something
changes between the base and the current image.

Remove some hooks that won't be run on the upgrade-base image
(which is only installed and then upgraded away from -- no expands
or shrinks currently).