[7.1/master] Remove Support for Debian 8 "Jessie" (#2210)
* Remove Debian 8 "Jessie" support from telekube & robotest apps.

Debian 8 has an old 3.16 linux kernel that does not have the overlay
module and has br_netfilter under the name bridge. We don't plan to test
on this version anymore.

* Format requirements table.

It is easier to read in plain text format this way.

* Remove Debian 8 "Jessie" from documented supported versions.

Due to Jessie's old 3.16 kernel, considerable modifications need to be
made to be able to run kubernetes.  Notably, Gravity 7.0.18 reports:

  [ERROR]: The following pre-flight checks failed:
        [×] kernel module "overlay" not loaded
        [×] kernel module "br_netfilter" not loaded
        [×] kernel module "overlay" not loaded
        [×] required kernel boot config parameter CONFIG_OVERLAY_FS missing
        [×] Following CGroups have not been mounted: ["memory"]
        [×] required kernel boot config parameter CONFIG_IP_NF_NA missing