lib/system*: introduce an attribute for explicit service state reload.
When creating a service unit, sometimes we repurpose the existing unit
files and the system manager (systemd) needs to be notified that the
existing unit file have been changed so it refreshes in-memory state.
Do this explicitly whenever a new service is created.

Updates https://github.com/gravitational/gravity/issues/2149.