[6.1.x] reload service definition before creating a service (#2257)
* lib/system*: introduce an attribute for explicit service state reload.

When creating a service unit, sometimes we repurpose the existing unit files and the system manager (systemd) needs to be notified that the existing unit file have been changed so it refreshes in-memory state. Do this explicitly whenever a new service is created.

Updates https://github.com/gravitational/gravity/issues/2149.

* lib/system/service: instead of stopping/disabling previous service, uninstall it when reinstalling a service.

This will make sure that the unit file is also removed when possible and at the correct moment in the whole flow.