[7.0.x] service state (#2240)
* Now that the install/join agents are running from a binary that resides in a known binary location (/usr/bin/ or /usr/local/bin) and local operation state is still kept in a working directory of tthe client, we need a way to let the service process know about the state directory.
Introduce a new CLI argument where necessary to convey this configuration and refactor handling of state for services.
* Explicit socket/service path configuration for the resume strategy.
* Remove excessive dependency on the state directory - instead provide functions with the information they would otherwise search in state directory directly.
* Address review comments
* Fix reconfiguration workflow (it is still missing service/pod subnet config though). Start self RPC commands using the current process's working directory

* lib/update/cluster: remove explicit package unpacking during bootstrap
step.

Bootstrap step will explicitly unpack all update packages into their
default location incl. the planet package. If the planet package did not
change, the installed and updated packages will be the same and the
unpacking will overwrite volatile container's configuration like the
resolv.conf.
Do not unpack the package to avoid overwriting. Figure out another way
to deal with https://github.com/gravitational/gravity/issues/1279 which
this change was introduced for.
* Download agent binary to /usr/local/bin by default in interactive installation to avoid collision with the installer on single node
* lib/system/service: instead of stopping/disabling previous service, uninstall it when reinstalling a service.
This will make sure that the unit file is also removed when possible and at the correct moment in the whole flow.
* Determine the correct binary to use for 'plan resume' depending on the operation type.