Increase dns-app update status checks
The number of retry attempts when checking the status of rigging changes may be too short for clusters experiencing high scheduler load. We have a customer app that appears to crash during the app update, creating lots of kubernetes scheduler events and the dns pods can take more than 2 minutes to start.