Conditionally create a user in the build container
We only want to create this user if they don't already exist.  This
allows the build to function when run as root (e.g. in a CI
environment). Without this the buildbox build fails with the following
error:

  useradd: UID 0 is not unique

(cherry picked from commit 0bed3091a2885e6c34fcbcf507cd93acdc18d730)