Do not use go modules when building tsh.
Tsh 3.2-g still uses Dep and vendoring to manage dependencies which
works when go does not try to re-download those but instead uses
the ones available in the repo.

294286adbac7bf36b63456809b7fc0938b7d95e4's introduction of Go 1.13
changed the behavior such that the teleport directory was getting
module-aware mode:

"The GO111MODULE environment variable continues to default to auto, but
the auto setting now activates the module-aware mode of the go command
whenever the current working directory contains, or is below a directory
containing, a go.mod file — even if the current directory is within
GOPATH/src."

Per https://golang.org/doc/go1.13#modules

The teleport directory lived under the gravity/build/teleport and
Gravity is module aware, so teleport was being flagged as such.

We don't test the build-tsh target as part of CI, so this was missed
until now.