Upgrade k8s client lib to 1.19.8 (#2433)
* Revendor

- Bump prometheus-operator to v0.40.1
- Bump k8s.io deps to v0.19.8
- Bump helm to v3.4.2

* Fix openapiv2 casing

* Provide context to k8s api

* Use InClusterConfig if path not provided

client-go complains that `Neither --kubeconfig nor --master was
specified...`, when getting kube client. This leads to failures in
robotest.  GetKubeClient has been updated to check if config is empty,
and if it is, then call InClusterConfig to get config.

GetKubeClientFromPath is redundant code and has been removed.

* Upgrade to helm v3

Rewrote helm actions. Ported
https://github.com/helm/helm/tree/v3.4.2/cmd/helm with some minor
tweaks.

Helm 3 no longer requires Tiller, and setup of directories and
repositories is automated

* Bump planet to 7.1.26

* Bump rigging to 7.1.3