Add UID and GID args to the buildbox
This was missed in 5e2b4c2e2bb1e1136a521e2c912060745f6b609a, resulting
in a buildbox without a UID corresponding to the UID of a developer
running the build as non-root.  This is mostly harmless, but will
cause anything that tries to pull user info in the buildbox to get
wierd results.