Fix ownership issues in buildbox
The buildbox build is split into two phases:

The first phase is performed as root -- and can write anywhere on the fs.
This is used to set up packages, as well as bootstrap directories for
use in the second phase.

The second phase operates as the local user, and sets up the remainder
of the gravity tool chain.

This backports structural parity with the post-gomodules dockerfile
introduced in 294286adbac7bf36b63456809b7fc0938b7d95e4.

The ownership issues arose from a careless backport in
6567fe26e2845aec19e1b17be08ba1e762b7eb47, which injected the local user
in place of jenkins, without also bringing back the above changes that
this depended on.

(cherry picked from commit e09ec65372b1d51ea2777612fb931fc75036fe58)