Fix ownership issues in buildbox
The buildbox build is split into two phases:

The first phase is performed as root -- and can write anywhere on the fs.
This is used to set up packages, as well as bootstrap directories for
use in the second phase.

The second phase operates as the local user, and sets up the remainder
of the gravity tool chain.

This backports structural parity with the post-gomodules dockerfile
introduced in 294286adbac7bf36b63456809b7fc0938b7d95e4.

The ownership issues arose from a careless backport in
5e2b4c2e2bb1e1136a521e2c912060745f6b609a, which injected the local user
in place of jenkins, without also bringing back the above changes that
this depended on.