[5.5] Add selinux and web/shared gitignore (#2477)
* Add selinux gitignore

These artifacts are only present in 7.0+, however because we
switch branches often, it is useful to git ignore these so that they
don't show up as potential changes.

* Backport gitignoring e/web/shared

This folder has ignored content in it in more recent gravity versions
that git notices when working on 5.5. Best to ignore it here too.